backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pillow==10.2.0
python-dotenv==1.0.1
transformers==4.38.2
torch==2.2.1
numpy==1.26.4
pydantic==2.6.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
cors==1.0.1